// frontend/src/entities/tree/model/tree.types.ts
export interface TreeState {
	data: string
	tree: {
		line: string
		treeLine: string
		imageUrl: string
	}
}

export enum TreeActionsTokens {
	UPDATE_TREE_TEXT_ACTION = 'UPDATE_TREE_TEXT_ACTION',

	CLEAR_TREE_PARSE_STATE_ACTION = 'CLEAR_TREE_PARSE_STATE_ACTION',

	ASYNC_PARSE_TREE_FROM_TEXT_ACTION = 'ASYNC_PARSE_TREE_FROM_TEXT_ACTION',

	LOADING_PARSE_TREE_FROM_TEXT_ACTION = 'LOADING_PARSE_TREE_FROM_TEXT_ACTION',
	LOADED_PARSE_TREE_FROM_TEXT_ACTION = 'LOADED_PARSE_TREE_FROM_TEXT_ACTION',
	ERROR_PARSE_TREE_FROM_TEXT_ACTION = 'ERROR_PARSE_TREE_FROM_TEXT_ACTION'
}